Support objects containing objects
The documentation for dart:mirrors is wrong, InstanceMirror.reflectee may be a
non-primative type, it can be an arbitary object.
In such cases we need to recursively transform to JSON as otherwise JSON.stringify dies later.
Here we first test if the object is of a primative type and if not we recurse.